Add support for ignored dependencies in lock state
This enables locking to be configured to ignore a set of module
identifiers. These identifiers will then be fully ignored from the
locking logic perspective: neither loaded nor written from/to lock
files, ignored when found in results. Note that these ignored module can
 also be totally absent from resolution.
The granularity of the set of ignored modules is per project.

Fixes #6846